added an entity builder
